--- Dockerfile-build
@@ -1,10 +1,0 @@
-FROM ec_golanggrpc_base
-
-ARG GROUP_ID
-ARG USER_ID
-
-RUN groupadd -g ${GROUP_ID} builder \
-  && useradd -d /home/builder -g builder -m -u ${USER_ID} builder \
-  && mkdir -p /home/builder/src
-
-WORKDIR /home/builder
