+++ buildscripts/README.md
@@ -10,7 +10,7 @@ separate buildscripts for each stage of the build process. Development can be do
 without needing to push upstream.
 
 Importantly, any suitable linux distro can be used as the development host provided that make, docker-ce,docker-compose,
-can be installed.
+are installed.
 
 
 #### Requirements
@@ -74,25 +74,23 @@ Be warned that this will require rebuilding the base container which can take > 
 ### List of make targets in priority order
 
 ```bash
-make remove_base          # deletes the docker base container
-make remove_container     # deletes the docker containers except base container
-make clean                # returns build environment to pristine state - does **not** revert any code changes
-make dependencies         # installs the dependencies
-make check                # statically checks code
-make unittest             # runs unittests
-make artifacts            # make the artifacts
-make                      # cleans and makes artifacts - execute this **before** pushing upstream to check errors
-make functest             # runs server in container and executes client from buildscripts/functest.sh
+make remove_base       # deletes the docker base container
+make remove_container  # deletes the docker containers except base container
+make clean             # returns build environment to pristine state - does **not** revert any code changes
+make generated         # create autogenerated code
+make check             # statically checks code
+make unittest          # runs unittests
+make artifacts         # make the artifacts
+make                   # cleans and makes artifacts - execute this **before** pushing upstream to check errors
+make functest          # runs server in container and executes client from buildscripts/functest.sh
 ```
-
-NB: Do not make any targets with the pattern _*_local. These are for internal use only.
 
 ### Debugging
 
 Before executing any make targets set an environment variable to 'DEBUG' to see log output:
 
 ```bash
-export EC_GOLANGGRPC_DEBUG="true"
+export GOLANGGRPC_DEBUG="true"
 make
 ```
 and log messages will be emitted.
