+++ buildscripts/clean.sh
@@ -3,5 +3,5 @@
 # returns repo to pristine state
 #
 set -e
+set -x
 git clean -fdX
-./buildscripts/env.sh
