--- buildscripts/dependencies.sh
@@ -1,9 +1,0 @@
-#!/bin/sh -e
-# 
-# get dependencies
-#
-cd src
-NAME=$( cat name  )
-
-# generate code from proto definition
-( cd ${NAME}/api && make )
