+++ buildscripts/docker-check
@@ -1,0 +1,21 @@
+#
+. ./buildscripts/env
+. ./buildscripts/log
+
+which docker-compose
+if [ $? -ne 0 ]
+then
+	error "docker-compose not installed. Exiting."
+fi
+docker-compose --version
+
+which docker
+if [ $? -ne 0 ]
+then
+	error "docker not installed. Exiting."
+fi
+docker --version
+
+# BUILDID is usually set by CI
+DOCKER_COMPOSE="docker-compose -p ${REPO}${BUILDID}"
+DOCKER="${DOCKER_COMPOSE} -f docker-compose.yaml"
