+++ cmd/client/main.go
@@ -7,7 +7,7 @@ import (
 
 	log "github.com/sirupsen/logrus"
 
-	"github.com/eccles/golanggrpc/ec_golanggrpc/client"
+	"github.com/eccles/golanggrpc/client"
 )
 
 func main() {
