+++ go.mod
@@ -3,8 +3,14 @@ module github.com/eccles/golanggrpc
 go 1.12
 
 require (
-	github.com/golang/protobuf v1.3.1
+	github.com/fullstorydev/grpcurl v1.3.2
+	github.com/go-delve/delve v1.2.0
+	github.com/golang/protobuf v1.3.2
+	github.com/grpc-ecosystem/grpc-gateway v1.9.6
+	github.com/rakyll/statik v0.1.6
 	github.com/sirupsen/logrus v1.4.0
+	github.com/spf13/cobra v0.0.5
 	golang.org/x/net v0.0.0-20190328230028-74de082e2cca
-	google.golang.org/grpc v1.19.1
+	golang.org/x/tools v0.0.0-20190311212946-11955173bddd
+	google.golang.org/grpc v1.21.0
 )
